MXM Provider: mq storage logic is moved to separate header
Apart from it more proper place for mq_storage_fini is chosen.
